chore: split HandleProtocolAdapter into different responsibilities.
