+++ src/main/java/edu/kit/datamanager/pit/pidsystem/impl/handle/HandleProtocolAdapter.java
@@ -1,16 +1,13 @@
 package edu.kit.datamanager.pit.pidsystem.impl.handle;
 
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
 import java.security.PrivateKey;
 import java.security.spec.InvalidKeySpecException;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.Map.Entry;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -28,17 +25,14 @@ import edu.kit.datamanager.pit.common.ExternalServiceException;
 import edu.kit.datamanager.pit.common.InvalidConfigException;
 import edu.kit.datamanager.pit.common.PidAlreadyExistsException;
 import edu.kit.datamanager.pit.common.PidNotFoundException;
-import edu.kit.datamanager.pit.common.PidUpdateException;
 import edu.kit.datamanager.pit.common.RecordValidationException;
 import edu.kit.datamanager.pit.configuration.HandleCredentials;
 import edu.kit.datamanager.pit.configuration.HandleProtocolProperties;
 import edu.kit.datamanager.pit.domain.PIDRecord;
-import edu.kit.datamanager.pit.domain.PIDRecordEntry;
 import edu.kit.datamanager.pit.pidsystem.IIdentifierSystem;
 import net.handle.api.HSAdapter;
 import net.handle.api.HSAdapterFactory;
 import net.handle.apps.batch.BatchUtil;
-import net.handle.hdllib.Common;
 import net.handle.hdllib.HandleException;
 import net.handle.hdllib.HandleResolver;
 import net.handle.hdllib.HandleValue;
@@ -55,30 +49,6 @@ import net.handle.hdllib.Util;
 public class HandleProtocolAdapter implements IIdentifierSystem {
 
     private static final Logger LOG = LoggerFactory.getLogger(HandleProtocolAdapter.class);
-
-    private static final byte[][][] BLACKLIST_NONTYPE_LISTS = {
-            Common.SITE_INFO_AND_SERVICE_HANDLE_INCL_PREFIX_TYPES,
-            Common.DERIVED_PREFIX_SITE_AND_SERVICE_HANDLE_TYPES,
-            Common.SERVICE_HANDLE_TYPES,
-            Common.LOCATION_AND_ADMIN_TYPES,
-            Common.SECRET_KEY_TYPES,
-            Common.PUBLIC_KEY_TYPES,
-            // Common.STD_TYPES, // not using because of URL and EMAIL
-            {
-                    // URL and EMAIL might contain valuable information and can be considered
-                    // non-technical.
-                    // Common.STD_TYPE_URL,
-                    // Common.STD_TYPE_EMAIL,
-                    Common.STD_TYPE_HSADMIN,
-                    Common.STD_TYPE_HSALIAS,
-                    Common.STD_TYPE_HSSITE,
-                    Common.STD_TYPE_HSSITE6,
-                    Common.STD_TYPE_HSSERV,
-                    Common.STD_TYPE_HSSECKEY,
-                    Common.STD_TYPE_HSPUBKEY,
-                    Common.STD_TYPE_HSVALLIST,
-            }
-    };
 
     private static final String SERVICE_NAME_HANDLE = "Handle System";
 
@@ -129,11 +99,10 @@ public class HandleProtocolAdapter implements IIdentifierSystem {
                     privateKey,
                     passphrase // "use null for unencrypted keys"
             );
-            HandleIndex indexManager = new HandleIndex();
             this.adminValue = this.client.createAdminValue(
                     props.getCredentials().getUserHandle(),
                     props.getCredentials().getPrivateKeyIndex(),
-                    indexManager.getHsAdminIndex());
+                    new HandleIndex().getHsAdminIndex());
         }
     }
 
@@ -168,9 +137,9 @@ public class HandleProtocolAdapter implements IIdentifierSystem {
             return null;
         }
         Collection<HandleValue> recordProperties = Streams.failableStream(allValues.stream())
-                .filter(value -> !isHandleInternalValue(value))
+                .filter(value -> !HandleBehavior.isHandleInternalValue(value))
                 .collect(Collectors.toList());
-        return new PIDRecord(recordProperties).withPID(pid);
+        return HandleBehavior.recordFrom(recordProperties).withPID(pid);
     }
 
     @NotNull
@@ -198,7 +167,7 @@ public class HandleProtocolAdapter implements IIdentifierSystem {
         for (RecordModifier modifier : this.props.getConfiguredModifiers()) {
             preparedRecord = modifier.apply(preparedRecord);
         }
-        ArrayList<HandleValue> futurePairs = handleValuesFrom(preparedRecord, Optional.of(admin));
+        ArrayList<HandleValue> futurePairs = HandleBehavior.handleValuesFrom(preparedRecord, Optional.of(admin));
 
         HandleValue[] futurePairsArray = futurePairs.toArray(new HandleValue[] {});
 
@@ -248,11 +217,11 @@ public class HandleProtocolAdapter implements IIdentifierSystem {
                 .collect(Collectors.toMap(HandleValue::getIndex, v -> v));
         // 1)
         List<HandleValue> valuesToKeep = oldHandleValues.stream()
-                .filter(HandleProtocolAdapter::isHandleInternalValue)
+                .filter(HandleBehavior::isHandleInternalValue)
                 .collect(Collectors.toList());
 
         // 2) Merge requested record and things we want to keep.
-        Map<Integer, HandleValue> recordNew = handleValuesFrom(preparedRecord, Optional.of(valuesToKeep))
+        Map<Integer, HandleValue> recordNew = HandleBehavior.handleValuesFrom(preparedRecord, Optional.of(valuesToKeep))
                 .stream()
                 .collect(Collectors.toMap(HandleValue::getIndex, v -> v));
 
@@ -354,69 +323,6 @@ public class HandleProtocolAdapter implements IIdentifierSystem {
     }
 
     /**
-     * Convert a `PIDRecord` instance to an array of `HandleValue`s. It is the
-     * inverse method to `pidRecordFrom`.
-     * 
-     * @param pidRecord the record containing values to convert / extract.
-     * @param toMerge   an optional list to merge the result with.
-     * @return HandleValues containing the same key-value pairs as the given record,
-     *         but e.g. without the name.
-     */
-    protected static ArrayList<HandleValue> handleValuesFrom(
-            final PIDRecord pidRecord,
-            final Optional<List<HandleValue>> toMerge)
-        {
-        ArrayList<Integer> skippingIndices = new ArrayList<>();
-        ArrayList<HandleValue> result = new ArrayList<>();
-        if (toMerge.isPresent()) {
-            for (HandleValue v : toMerge.get()) {
-                result.add(v);
-                skippingIndices.add(v.getIndex());
-            }
-        }
-        HandleIndex index = new HandleIndex().skipping(skippingIndices);
-        Map<String, List<PIDRecordEntry>> entries = pidRecord.getEntries();
-
-        for (Entry<String, List<PIDRecordEntry>> entry : entries.entrySet()) {
-            for (PIDRecordEntry val : entry.getValue()) {
-                String key = val.getKey();
-                HandleValue hv = new HandleValue();
-                int i = index.nextIndex();
-                hv.setIndex(i);
-                hv.setType(key.getBytes(StandardCharsets.UTF_8));
-                hv.setData(val.getValue().getBytes(StandardCharsets.UTF_8));
-                result.add(hv);
-                LOG.debug("Entry: ({}) {} <-> {}", i, key, val);
-            }
-        }
-        assert result.size() >= pidRecord.getEntries().size();
-        return result;
-    }
-
-    protected static class HandleIndex {
-        // handle record indices start at 1
-        private int index = 1;
-        private List<Integer> skipping = new ArrayList<>();
-
-        public final int nextIndex() {
-            int result = index;
-            do {
-                index += 1;
-            } while (index == this.getHsAdminIndex() || skipping.contains(index));
-            return result;
-        }
-
-        public HandleIndex skipping(List<Integer> skipThose) {
-            this.skipping = skipThose;
-            return this;
-        }
-
-        public final int getHsAdminIndex() {
-            return 100;
-        }
-    }
-
-    /**
      * Returns true if the PID is valid according to the following criteria:
      * - PID is valid according to isIdentifierRegistered
      * - If a generator prefix is set, the PID is expedted to have this prefix.
@@ -438,97 +344,5 @@ public class HandleProtocolAdapter implements IIdentifierSystem {
             return false;
         }
         return true;
-    }
-
-    /**
-     * Checks if a given value is considered an "internal" or "handle-native" value.
-     * <p>
-     * This may be used to filter out administrative information from a PID record.
-     * 
-     * @param v the value to check.
-     * @return true, if the value is conidered "handle-native".
-     */
-    public static boolean isHandleInternalValue(HandleValue v) {
-        boolean isInternalValue = false;
-        for (byte[][] typeList : BLACKLIST_NONTYPE_LISTS) {
-            for (byte[] typeCode : typeList) {
-                isInternalValue = isInternalValue || Arrays.equals(v.getType(), typeCode);
-            }
-        }
-        return isInternalValue;
-    }
-
-    /**
-     * Given two Value Maps, it splits the values in those which have been added,
-     * updated or removed.
-     * Using this lists, an update can be applied to the old record, to bring it to
-     * the state of the new record.
-     */
-    protected static class HandleDiff {
-        private final Collection<HandleValue> toAdd = new ArrayList<>();
-        private final Collection<HandleValue> toUpdate = new ArrayList<>();
-        private final Collection<HandleValue> toRemove = new ArrayList<>();
-
-        HandleDiff(
-            final Map<Integer, HandleValue> recordOld,
-            final Map<Integer, HandleValue> recordNew
-        ) throws PidUpdateException {
-            for (Entry<Integer, HandleValue> old : recordOld.entrySet()) {
-                boolean wasRemoved = !recordNew.containsKey(old.getKey());
-                if (wasRemoved) {
-                    // if a row in the record is not available anymore, we need to delete it
-                    toRemove.add(old.getValue());
-                } else {
-                    // otherwise, we should go and update it.
-                    // we could also check for equality, but this is the safe and easy way.
-                    // (the handlevalue classes can be complicated and we'd have to check their
-                    // equality implementation)
-                    toUpdate.add(recordNew.get(old.getKey()));
-                }
-            }
-            for (Entry<Integer, HandleValue> e : recordNew.entrySet()) {
-                boolean isNew = !recordOld.containsKey(e.getKey());
-                if (isNew) {
-                    // if there is a record which is not in the oldRecord, we need to add it.
-                    toAdd.add(e.getValue());
-                }
-            }
-
-            // runtime testing to avoid messing up record states.
-            String exceptionMsg = "DIFF NOT VALID. Type: %s. Value: %s";
-            for (HandleValue v : toRemove) {
-                boolean valid = recordOld.containsValue(v) && !recordNew.containsKey(v.getIndex());
-                if (!valid) {
-                    String message = String.format(exceptionMsg, "Remove", v.toString());
-                    throw new PidUpdateException(message);
-                }
-            }
-            for (HandleValue v : toAdd) {
-                boolean valid = !recordOld.containsKey(v.getIndex()) && recordNew.containsValue(v);
-                if (!valid) {
-                    String message = String.format(exceptionMsg, "Add", v);
-                    throw new PidUpdateException(message);
-                }
-            }
-            for (HandleValue v : toUpdate) {
-                boolean valid = recordOld.containsKey(v.getIndex()) && recordNew.containsValue(v);
-                if (!valid) {
-                    String message = String.format(exceptionMsg, "Update", v);
-                    throw new PidUpdateException(message);
-                }
-            }
-        }
-
-        public HandleValue[] added() {
-            return this.toAdd.toArray(new HandleValue[] {});
-        }
-
-        public HandleValue[] updated() {
-            return this.toUpdate.toArray(new HandleValue[] {});
-        }
-
-        public HandleValue[] removed() {
-            return this.toRemove.toArray(new HandleValue[] {});
-        }
     }
 }
